+++ python_api_testing/models/bert/add_and_norm.py
@@ -57,9 +57,9 @@ def run_add_and_norm_inference():
     pytorch_out = pytorch_add_and_norm_model(add_and_norm_inputa, add_and_norm_inputb)
 
     tt_add_and_norm_input_a = tilize_to_list(pad_activation(add_and_norm_inputa))
-    tt_add_and_norm_input_a = gpai.tensor.Tensor(tt_add_and_norm_input_a, add_and_norm_inputa.shape, gpai.tensor.DataType.FLOAT32,  gpai.tensor.Layout.TILE, device)
+    tt_add_and_norm_input_a = gpai.tensor.Tensor(tt_add_and_norm_input_a, add_and_norm_inputa.shape, gpai.tensor.DataType.BFLOAT16,  gpai.tensor.Layout.TILE, device)
     tt_add_and_norm_input_b = tilize_to_list(pad_activation(add_and_norm_inputb))
-    tt_add_and_norm_input_b = gpai.tensor.Tensor(tt_add_and_norm_input_b, add_and_norm_inputb.shape, gpai.tensor.DataType.FLOAT32,  gpai.tensor.Layout.TILE, device)
+    tt_add_and_norm_input_b = gpai.tensor.Tensor(tt_add_and_norm_input_b, add_and_norm_inputb.shape, gpai.tensor.DataType.BFLOAT16,  gpai.tensor.Layout.TILE, device)
 
     tt_out = tt_add_and_norm_model(tt_add_and_norm_input_a, tt_add_and_norm_input_b).to(host)
     tt_out = untilize(torch.Tensor(tt_out.data()).reshape(*pytorch_out.shape))
